general cleanup, fix selector
The current selector was going through all rows, not only visible
ones. While debugging this I have realized that loading one of the
sports pages is enough to parse all leagues for all sports, because
all of them are right there in the HTML, they are just not visible.

In the future that will speed things up a bit, but for the time being
let's leave it as it is.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>