WebData: introduce class to store extra data
Introducing this intermediate class allows to implement,
in the future, the following:

* Carry additional data: this will be useful to get rid
of the ugly Pair<> return typo on the getTabs() method.

* Add debug data: similar to the previous point, but this
could be disabled by some compilation constant. For example
adding a screenshot of each page to this class would allow
better diagnostic information when dumping ScrapExceptions.

* Lazy parsing: The Jsoup.parse() call could be invoked on
the first getDoc() call, rather than when constructing the
object. This means the parsing effort will automatically be
done by the parsing class, not the html provider. I plan to
make this application multithreaded later and the html
provider will most likely be the bottleneck.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>