StringDate: implement actual parsing
At least add the format that is being reported in most of the odds right
now. I will keep adding formats as they appear. Hopefully there are no
ambiguities, but I can already forecast some problems when parsing dates
for the next year (in late december) as the year is not included in the
DateString...

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>