OddsPortalScrapper: do not log errors if a League is truly empty
Some times the web displays leagues even though when you click of them
there is no matches to be parsed. In those cases the web shows an info box
about that. If there are no matches to be parsed and the info box is there
we can be sure that there were no errors parsing: there are simply no
matches.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>