SQLiteManager_v2: fix deadlock in close()
If the working thread is waiting on BlockingQueue::take() (which is quite
likely) and a thread calls SQLiteManager_v2::close() (which calls
WorkingThread::stopAndWait()) this would trigger a deadlock, as the
WorkingThread will never have a chance to check the 'running' flag.

Use an special value in the work queue to signal finish. As of right now
this is not a perfect solution and it has at least one race condition, but
it is a good enough solution for now.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>