Utils: PriorityExecutor: implement an executor that supports priorities
I find it quite surprising that something like this is not available in
standard Java. This is a simple implementation that allows to submit tasks
with different priorities.

When I make this application I will want to be able to query things like
parsing live matches when higher priority that parsing other things (as
live odds change quicker and contains more information about the event).

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>